Add configurable PPO module and config loading
