+++ tests/test_config_loading.py
@@ -1,0 +1,14 @@
+import os
+import unittest
+
+from ppo import load_config
+
+class TestConfigLoading(unittest.TestCase):
+    def test_json_config(self):
+        path = os.path.join(os.path.dirname(__file__), 'sample_config.json')
+        cfg = load_config(path)
+        self.assertEqual(cfg['ppo']['batch_size'], 32)
+        self.assertEqual(cfg['curriculum']['threshold'], 0.5)
+
+if __name__ == '__main__':
+    unittest.main()
